% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_ncell}
\alias{grid_ncell}
\title{Number of cells in a grid}
\usage{
grid_ncell(grid)
}
\arguments{
\item{grid}{a \code{guerrilla_grid}}
}
\value{
A single integer, the product of the two dimensions.
}
\description{
Number of cells in a grid
}
\examples{
grid_ncell(grid_spec(dimension = c(3, 2), extent = c(0, 6, 0, 4)))
}
