% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_res}
\alias{grid_res}
\title{Cell size in each direction}
\usage{
grid_res(grid)
}
\arguments{
\item{grid}{a \code{guerrilla_grid}}
}
\value{
Two numbers, the x and y size of one cell.
}
\description{
Cell size in each direction
}
\examples{
grid_res(grid_spec(dimension = c(4, 2), extent = c(0, 8, 0, 4)))
}
