% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{is_grid}
\alias{is_grid}
\title{Is this a guerrilla grid?}
\usage{
is_grid(x)
}
\arguments{
\item{x}{any object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Is this a guerrilla grid?
}
\examples{
is_grid(grid_spec(cbind(1:3, 1:3)))
is_grid(1:10)
}
