% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{plot.guerrilla_grid}
\alias{plot.guerrilla_grid}
\title{Plot a grid}
\usage{
\method{plot}{guerrilla_grid}(x, col = grDevices::hcl.colors(24, "YlGnBu"), asp = 1, ...)
}
\arguments{
\item{x}{a \code{guerrilla_grid} with values}

\item{col}{colours}

\item{asp}{aspect ratio; \code{1} by default, or use \code{NA} to fill the device}

\item{...}{passed to \code{\link[graphics:image]{graphics::image()}}}
}
\value{
\code{x}, invisibly.
}
\description{
Draw the grid with \code{\link[graphics:image]{graphics::image()}}. No raster class is involved, which
is the point: a grid is a dimension, an extent and a vector of values, and
that is enough to draw it.
}
\examples{
xy <- cbind(runif(50), runif(50))
g <- grid_barycentric(xy, xy[, 1] + xy[, 2])
plot(g)
}
