% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilinear_triangulation.R
\name{tri_pip}
\alias{tri_pip}
\title{Identify point-in-triangle by conversion to polygons}
\usage{
tri_pip(tri, pts)
}
\arguments{
\item{tri}{list P n\emph{2 coordinates and T matrix of n}3 indices defining triangles}

\item{pts}{input points}
}
\description{
Identify point-in-triangle by conversion to polygons
}
