% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{converters}
\alias{converters}
\alias{as_raster}
\alias{as_terra}
\alias{as_gdalraster}
\title{Hand a grid to another package}
\usage{
as_raster(grid)

as_terra(grid)

as_gdalraster(grid, filename = NULL)
}
\arguments{
\item{grid}{a \code{guerrilla_grid}}

\item{filename}{file to write; defaults to a path in GDAL's \verb{/vsimem}}
}
\value{
An object of the corresponding class.
}
\description{
Converters to the raster classes people actually have installed. Each one is
a handful of lines, because there is nothing in a grid that these classes do
not also hold: the same dimension, the same extent, the same values in the
same order.
}
\details{
\code{as_gdalraster()} has to write a file, since that is how GDAL works; by
default it writes to GDAL's in-memory filesystem.
}
\examples{
xy <- cbind(runif(50), runif(50))
g <- tri_fun(xy, xy[, 1])
if (requireNamespace("terra", quietly = TRUE)) as_terra(g)
}
