% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defaultgrid}
\alias{defaultgrid}
\title{Title}
\usage{
defaultgrid(xy, ncol = 60, nrow = 50,
  prj = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0")
}
\arguments{
\item{xy}{coordinates}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{prj}{projection metadata}
}
\value{
raster
}
\description{
Title
}
