% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybetin.R
\name{facets}
\alias{facets}
\title{Title}
\usage{
facets(
  X,
  nx,
  ny,
  x = NULL,
  y = NULL,
  na.v = 0,
  method = c("dirichlet", "delaunay")
)
}
\arguments{
\item{X}{spatstat object}

\item{nx}{number of x coords}

\item{ny}{number of y coords}

\item{x}{option input x values}

\item{y}{optional input y values}

\item{na.v}{na value}

\item{method}{dirichlet or delaunay}
}
\value{
ppp object
}
\description{
Title
}
