% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg-package.R
\docType{package}
\name{guerrilla-package}
\alias{guerrilla}
\alias{guerrilla-package}
\title{guerrilla: Illustrate Various Methods of Interpolation for Irregular Data}
\description{
Examples of interpolating irregular data, to illustrate the 
 mechanics of various methods and some easy tools to run them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/guerrilla}
  \item Report bugs at \url{https://github.com/hypertidy/guerrilla/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Marcelino de la Cruz (wrote maybetin example function on R-Sig-Geo) [contributor]
  \item Tomas Remenyi [contributor]
}

}
\keyword{internal}
