% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg-package.R
\docType{package}
\name{irreg-package}
\alias{irreg}
\alias{irreg-package}
\title{irreg: Illustrate Various Methods of Interpolation for Irregular Data}
\description{
Examples of interpolating irregular data, to illustrate the mechanics
of various methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/irreg}
  \item Report bugs at \url{https://github.com/mdsumner/irreg/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Marcelino de la Cruz [contributor]
  \item Tomas Remenyi [contributor]
}

}
\keyword{internal}
