% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_raster.R
\name{mesh_raster}
\alias{mesh_raster}
\alias{mesh_raster.mesh3d}
\alias{mesh_raster.matrix}
\alias{mesh_raster.data.frame}
\title{Mesh raster}
\usage{
mesh_raster(x, grid = NULL, n = 128)

\method{mesh_raster}{mesh3d}(x, grid = NULL, n = 128)

\method{mesh_raster}{matrix}(x, grid = NULL, n = 128)

\method{mesh_raster}{data.frame}(x, grid = NULL, n = 128)
}
\arguments{
\item{x}{matrix of points, or a mesh3d}

\item{grid}{raster to populate}

\item{n}{grid size of raster if 'grid' not supplied}
}
\value{
Raster
}
\description{
Create a raster by interpolating across triangles
}
\details{
Barycentric interpolation is used to efficiently obtain
a within-triangle estimate of a field of values
}
\examples{
data("humface", package = "Rvcg")
x <- humface
grid <- mesh_raster(x, n = 256)
raster::plot(grid, col = grey.colors(21), 
    breaks = quantile(grid, seq(0, 1, length = 22), na.rm = TRUE))
anglr::plot3d(grid)
}
