% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilinear_triangulation.R
\name{tri_fun}
\alias{tri_fun}
\title{Title}
\usage{
tri_fun(xy, value, grid = NULL, ...)
}
\arguments{
\item{xy}{coordinates}

\item{value}{value to interpolate}

\item{grid}{grid to use}

\item{...}{ignored}
}
\value{
raster
}
\description{
Title
}
